EidUpdateSendFormRequestDto moved to form sender and nases type guards extracted
